old projekt/Ny remote1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.0
python-socketio==5.8.0
aiortc>=1.6.0
av>=11.0.0
PyQt6==6.6.0
opencv-python==4.8.1.78
numpy==1.26.2
python-multipart==0.0.6
websockets==12.0
pillow==10.1.0
cython==0.29.37
setuptools==68.2.2
Flask==3.0.0
redis==5.0.0
flask-cors==4.0.0
